Anise Hyssop/anise-hyssop-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.36.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
